move NavPvtFlags and NavPvtFlags2 into common module for NAV-PVT to eliminate duplicate code
