+++ ublox/src/ubx_packets/packets/nav_pvt/common.rs
@@ -1,0 +1,38 @@
+use bitflags::bitflags;
+use ublox_derive::ubx_extend_bitflags;
+
+#[ubx_extend_bitflags]
+#[ubx(from, rest_reserved)]
+bitflags! {
+    /// Fix status flags for `NavPvt`
+    #[derive(Debug)]
+    pub struct NavPvtFlags: u8 {
+        /// Position and velocity valid and within DOP and ACC Masks
+        const GPS_FIX_OK = 1;
+        /// Differential corrections were applied; DGPS used
+        const DIFF_SOLN = 2;
+        /// Heading of vehicle is valid
+        const HEAD_VEH_VALID = 0x20;
+        const CARR_SOLN_FLOAT = 0x40;
+        const CARR_SOLN_FIXED = 0x80;
+    }
+}
+
+#[ubx_extend_bitflags]
+#[ubx(from, rest_reserved)]
+bitflags! {
+    /// Additional flags for `NavPvt`
+    #[derive(Debug)]
+    pub struct NavPvtFlags2: u8 {
+        /// 1 = information about UTC Date and Time of Day validity confirmation
+        /// is available. This flag is only supported in Protocol Versions
+        /// 19.00, 19.10, 20.10, 20.20, 20.30, 22.00, 23.00, 23.01,27 and 28.
+        const CONFIRMED_AVAI = 0x20;
+        /// 1 = UTC Date validity could be confirmed
+        /// (confirmed by using an additional independent source)
+        const CONFIRMED_DATE = 0x40;
+        /// 1 = UTC Time of Day could be confirmed
+        /// (confirmed by using an additional independent source)
+        const CONFIRMED_TIME = 0x80;
+    }
+}
