+++ ublox/src/ubx_packets/packets/nav_pvt/proto14.rs
@@ -5,6 +5,7 @@ use super::SerializeUbxPacketFields;
 #[cfg(feature = "serde")]
 use crate::serde::ser::SerializeMap;
 
+use super::common::*;
 use crate::{error::ParserError, GnssFixType, UbxPacketMeta};
 use ublox_derive::{ubx_extend_bitflags, ubx_packet_recv};
 
@@ -96,74 +97,4 @@ struct NavPvt {
 
     reserved2: [u8; 2],
     reserved3: [u8; 4],
-}
-
-#[ubx_extend_bitflags]
-#[ubx(from, rest_reserved)]
-bitflags! {
-    /// Fix status flags for `NavPvt`
-    #[derive(Debug)]
-    pub struct NavPvtFlags: u8 {
-        /// Position and velocity valid and within DOP and ACC Masks
-        const GPS_FIX_OK = 1;
-        /// Differential corrections were applied; DGPS used
-        const DIFF_SOLN = 2;
-        /// Heading of vehicle is valid
-        const HEAD_VEH_VALID = 0x20;
-        const CARR_SOLN_FLOAT = 0x40;
-        const CARR_SOLN_FIXED = 0x80;
-    }
-}
-
-#[ubx_extend_bitflags]
-#[ubx(from, rest_reserved)]
-bitflags! {
-    /// Additional flags for `NavPvt`
-    #[derive(Debug)]
-    pub struct NavPvtFlags2: u8 {
-        /// 1 = information about UTC Date and Time of Day validity confirmation
-        /// is available. This flag is only supported in Protocol Versions
-        /// 19.00, 19.10, 20.10, 20.20, 20.30, 22.00, 23.00, 23.01,27 and 28.
-        const CONFIRMED_AVAI = 0x20;
-        /// 1 = UTC Date validity could be confirmed
-        /// (confirmed by using an additional independent source)
-        const CONFIRMED_DATE = 0x40;
-        /// 1 = UTC Time of Day could be confirmed
-        /// (confirmed by using an additional independent source)
-        const CONFIRMED_TIME = 0x80;
-    }
-}
-
-#[cfg(feature = "ubx_proto23")]
-pub(crate) mod flags {
-    #[derive(Debug, Clone, Copy)]
-    #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-    pub struct NavPvtFlags3 {
-        invalid_llh: bool,
-        age_differential_correction: u8,
-    }
-
-    impl NavPvtFlags3 {
-        pub fn invalid_llh(&self) -> bool {
-            self.invalid_llh
-        }
-
-        pub fn age_differential_correction(&self) -> u8 {
-            self.age_differential_correction
-        }
-    }
-
-    impl From<u8> for NavPvtFlags3 {
-        fn from(val: u8) -> Self {
-            const AGE_DIFFERENTIAL_CORRECTION_MASK: u8 = 0b11110;
-            let invalid = val & 0x01 == 1;
-            // F9R interface description document specifies that this byte is unused
-            // We can read it ... but we don't expose it
-            let age_differential_correction = val & AGE_DIFFERENTIAL_CORRECTION_MASK;
-            Self {
-                invalid_llh: invalid,
-                age_differential_correction,
-            }
-        }
-    }
 }
