+++ ublox/src/ubx_packets/packets/nav_pvt/proto23_27_31.rs
@@ -1,12 +1,11 @@
-use bitflags::bitflags;
-
 #[cfg(feature = "serde")]
 use super::super::SerializeUbxPacketFields;
 #[cfg(feature = "serde")]
 use crate::serde::ser::SerializeMap;
 
+use super::common::*;
 use crate::{error::ParserError, GnssFixType, UbxPacketMeta};
-use ublox_derive::{ubx_extend_bitflags, ubx_packet_recv};
+use ublox_derive::ubx_packet_recv;
 
 /// Navigation Position Velocity Time Solution
 #[ubx_packet_recv]
@@ -107,42 +106,6 @@ struct NavPvt {
 
     #[ubx(map_type = f64, scale = 1e-2, alias = magnetic_declination_accuracy)]
     magnetic_declination_accuracy: u16,
-}
-
-#[ubx_extend_bitflags]
-#[ubx(from, rest_reserved)]
-bitflags! {
-    /// Fix status flags for `NavPvt`
-    #[derive(Debug)]
-    pub struct NavPvtFlags: u8 {
-        /// Position and velocity valid and within DOP and ACC Masks
-        const GPS_FIX_OK = 1;
-        /// Differential corrections were applied; DGPS used
-        const DIFF_SOLN = 2;
-        /// Heading of vehicle is valid
-        const HEAD_VEH_VALID = 0x20;
-        const CARR_SOLN_FLOAT = 0x40;
-        const CARR_SOLN_FIXED = 0x80;
-    }
-}
-
-#[ubx_extend_bitflags]
-#[ubx(from, rest_reserved)]
-bitflags! {
-    /// Additional flags for `NavPvt`
-    #[derive(Debug)]
-    pub struct NavPvtFlags2: u8 {
-        /// 1 = information about UTC Date and Time of Day validity confirmation
-        /// is available. This flag is only supported in Protocol Versions
-        /// 19.00, 19.10, 20.10, 20.20, 20.30, 22.00, 23.00, 23.01,27 and 28.
-        const CONFIRMED_AVAI = 0x20;
-        /// 1 = UTC Date validity could be confirmed
-        /// (confirmed by using an additional independent source)
-        const CONFIRMED_DATE = 0x40;
-        /// 1 = UTC Time of Day could be confirmed
-        /// (confirmed by using an additional independent source)
-        const CONFIRMED_TIME = 0x80;
-    }
 }
 
 #[cfg(feature = "ubx_proto23")]
